Yang Yao/requirements.txt
streamlit>=1.32
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

pandas
numpy
yfinance
matplotlib
reportlab>=4.0.0

